+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed a nullref when a custom post process component that was in the HDRP PP list is removed from the project
 - Fixed issue that prevented decals from modifying specular occlusion (case 1178272).
 - Fixed exposure of volumetric reprojection
+- Fixed multi selection support for Scalable Settings in lights
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled